requirements.txt
matplotlib
numpy
scipy
python-dotenv
PyQt5